requirements.txt
aniso8601==7.0.0
asn1crypto==0.24.0
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
cryptography==2.7
Flask==1.1.0
Flask-Bcrypt==0.7.1
flask-marshmallow==0.10.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-sqlalchemy==0.17.0
psycopg2-binary==2.8.3
pycparser==2.19
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.5
Werkzeug==0.15.4
